moved to the correct directory
moved to the correct directory (and also trying out github)
